automatically backed up by learn

+++ lib/bachelor.rb
@@ -43,7 +43,7 @@ def get_average_age_for_season(data, season)
   contestants = data[season]
   sum = 0.0
   contestants.each do |attrs|
-    sum += attrs["age"].to_i
+    sum += attrs["age"].to_f
   end
   sum / contestants.length 
 end
